altering type of column price to double

+++ backend/src/main/java/com/proghuly/cardapio/DTO/FoodRequestDTO.java
@@ -1,5 +1,5 @@
 package com.proghuly.cardapio.DTO;
 
-public record FoodRequestDTO(String title, String image, Integer price) {
+public record FoodRequestDTO(String title, String image, double price) {
 
 }
